pyproject.toml
[project]
name = "new-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-tailwind>=3.8.0",
]
